backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
loguru==0.7.2
openai==1.51.2
httpx>=0.27.2
anyio>=4.4.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3